backend fixes

+++ backend/app/utils.py
@@ -24,11 +24,11 @@ def preprocess_image(image: Image.Image) -> np.ndarray:
     image_array = np.transpose(image_array, (2, 0, 1))
     
     # Normalize with mean 0.5, std 0.5
-    mean = np.array([0.5, 0.5, 0.5]).reshape(-1, 1, 1)
-    std = np.array([0.5, 0.5, 0.5]).reshape(-1, 1, 1)
+    mean = np.array([0.5, 0.5, 0.5], dtype=np.float32).reshape(-1, 1, 1)
+    std = np.array([0.5, 0.5, 0.5], dtype=np.float32).reshape(-1, 1, 1)
     image_array = (image_array - mean) / std
     
-    # Add batch dimension
-    input_numpy = image_array[np.newaxis, ...]
+    # Add batch dimension and ensure float32
+    input_numpy = image_array[np.newaxis, ...].astype(np.float32)
     
     return input_numpy
